Smaller ints

+++ maze.c
@@ -13,7 +13,7 @@ void maze_init(maze *maze) {
   maze->nodes = NULL;
 }
 
-node* maze_node_at(maze *maze, int64_t x, int64_t y) {
+node* maze_node_at(maze *maze, int32_t x, int32_t y) {
   if (x < 0 || x >= maze->width) return NULL;
   if (y < 0 || y >= maze->height) return NULL;
   return maze->nodes[y * maze->width + x];
@@ -24,7 +24,7 @@ void maze_read(maze *maze, FILE *stream) {
     char *line = NULL;
     size_t len = 0, i;
     ssize_t read = getline(&line, &len, stream);
-    uint64_t y;
+    uint32_t y;
 
     if (read == -1) {
       free(line);
@@ -94,12 +94,12 @@ node* maze_goal(maze *maze) {
   return maze_node_at(maze, maze->gx, maze->gy);
 }
 
-void _check_and_add_node(maze *maze, nodeset *set, int64_t x, int64_t y) {
+void _check_and_add_node(maze *maze, nodeset *set, int32_t x, int32_t y) {
   node *node = maze_node_at(maze, x, y);
   if (node != NULL) nodeset_add(set, node);
 }
 
-void maze_neighbors(maze *maze, int64_t x, int64_t y, nodeset* output) {
+void maze_neighbors(maze *maze, int32_t x, int32_t y, nodeset* output) {
   nodeset_clear(output);
 
   _check_and_add_node(maze, output, x - 1, y);
@@ -108,7 +108,7 @@ void maze_neighbors(maze *maze, int64_t x, int64_t y, nodeset* output) {
   _check_and_add_node(maze, output, x, y + 1);
 }
 
-uint64_t maze_heuristic(maze *maze, node *node) {
+uint32_t maze_heuristic(maze *maze, node *node) {
   return abs(node->x - maze->gx) + abs(node->y - maze->gy);
 }
 
@@ -132,7 +132,7 @@ void maze_astar(maze *maze) {
   while (open.size > 0) {
     size_t i;
     node *current = open.nodes[0];
-    uint64_t best = current->fscore;
+    uint32_t best = current->fscore;
 
     for (i = 1; i < open.size; ++i) {
       if (open.nodes[i]->fscore <= best) {
@@ -155,7 +155,7 @@ void maze_astar(maze *maze) {
 
     for (i = 0; i < neighbors.size; ++i) {
       node *neighbor = neighbors.nodes[i];
-      uint64_t gscore = current->gscore + 1;
+      uint32_t gscore = current->gscore + 1;
 
       if (nodeset_contains(&closed, neighbor)) continue;
